requirements/base.txt
django
gevent-socketio
django-tastypie
django-simple-audit